fileUtil: add loadOffsets()
Reads a file "offsets.txt" in dirName that is a file that contains x,y
pairs matching each image in the directory. For example, and a directory
with 3 images must have an offets.txt file with an offset for each
image:

offsets.txt {
  x y
  x y
  x y
}

Note that files are sorted in lexicographic order.