EncodeManager now records VNC session to file
Dumps the entire framebuffer each update and writes it to file using QOI
encoding. Right now, it seems as if the client is the bottleneck and not
the server (neither CPU or IO are maxed out). The frames are only dumped
when the client requests it, and it doesn't seem to do so fast enough.
I'm looking into moving this code higher up so that a client isn't
needed to record a session.