Add file metadata to Frame(Out|In)Stream
The header tells which encoding was used, and the framebuffer size when
recording. The structure is a bit messy, using primitive reflections to
construct a decoder during runtime when the file is parsed. It's a bit
messy to use as well, as we need to ensure that the header is
parsed/written first, and that the flags are set correctly.