FrameOutStream: remove unused buffer
It seems like it isn't worth saving a bunch of writes in memory in
larger batches, as a single image is large enough to not warrant any
overhead/performance losses. I think one way that might improve
performance would be to parallellize the encoding, but keep writes
ordered. This should work if we're CPU bound and not I/O bound (as the
case seems to be right now).