Recoder: only write update on DAMAGE event
Uses X11 DAMAGE extension to detect changed to framebuffer instead of
polling in a busy loop. This saves filesizes dramatically when not a lot
of things are happening on screen, & increases performance as well. So
instead of saving entire screen updates, only parts of the screen that
have chagned will be saved. This needs more experimentation to see if
this is viable. It seemes to work, but there might be differences in how
applications use this extensions, which can have some side effects.
Also, instead of re-imlpementing parts of XDesktop, I want to look if I
can use the classes directly, and only inherit/override small parts to
make it so that FileOutStream is used instead of actually encoding to
RFB OutStream.