Fix how compression ratio is calculated
Previously, the compression ratio was looking at the ratio of the
recorder/input file, and not the ratio of the Server's EncodeManager.
This was fixed to look at the ratio between the raw image (w*h*3)
divided by the Server's outputstream, meaning we get the compression
ratio of the RFB protocol as a whole. We might want to distinguish
between the RFB ratio and only the encoding ratio later on.