Recorder.cxx: coalesce all DAMAGE events
Whenver a DAMAGE event is detected, we coalesce all DAMAGE events into a
bounding rectangle and write one update. Since DAMAGE events only
contain the area that was damaged, and not the contents of the
framebuffer when the damage occured, overlaps in DAMAGE events will
cause conflicts in the image. Instead, we create a larger, bounding
rectangle around all damaged regions and save them as one update. This
removes inconsistencies when these interesctions occurs. It seems to
work fine right now, but probably needs more testing.