Benchmark now prints encoders timing statistics
We now keep track of the time it takes for:
  * Encoders to run writeRect() & writeSolidRect()
  * EncodeManager to run writeUpdate (RFB + encoding)
  * Time spent actually encoding images vs. RFB/EncodeManager stuff
    (spliting images, palette analysis etc).

We also keep track of the compresion ratio going from raw image to data
on the RFB protocol stream. This needs to be complemented with keeping
track of only the image compression ratio. A caveat with this is that we
also probably want separate how we calculate solidRects compression from
normal rect compression.