Add compression stats to TimedEncoders
An ugly way of getting the amount of bytes written by writeRect() and
writeSolidRect(). It works by "injecting" our own OutStream to the
underlying SConnection's OutStream, so that the encoder writes all data
to our OutStream when encoding. After a rectangle is encoded, we put
the original OutStream back so that we can use it to measure RFB
protocol overhead (if we want to).

Everytime our OutStream is injected, we clear it when we're done so that
ptr = start, without needing to do any memory reallocations. This
shouldn't affect performance too much.