Add constructor to Manager that takes Encoder[].
This way, we can more precisely control which encoders to test. Instead
of using setEncodings() only, we can overwrite the encoders[] array in
EncodeManager and decide which encoder should be at any given index
(0-5). This means we can eliminate instances where a given encoding is
set, such as RRE for example, and where EncodeManager decides to use a
different encoder for specific rectangles. This is useful if we want to
only test specific encoders, and not the EncodeManager.