Keep track of frametime data
We want to keep track of how long each frame update takes to encode, and
the time budget before the next frame occurs in the recorded file. This
way, we can take measurements of how good/bad the EncodeManager performs
in relation to the input file (assumes we want to play it back in
realtime).