Split encoderStats into ManagerStats/EncoderStats
encoderstats.h was getting a bit confusing for me to use and was in need
of refactoring anyways. Thought it might be best to separate between the
EncodeManager stats (which include RFB overhead), and only the Encoder
stats. This change is needed to make it possible for a Manager to use
different Encoders & track the statistics accordingly. I don't like the
signature change for TimedEncoder, which now  needs a pointer to
Manager* so that it can tell it which Encoder was used for which
writeUpdate.