Recorder updates at set interval
Users can now pass a framerate as an argument, and the Recorder will
attempt to only process updates at that interval. I think doing it this
way, simulating a normal screen recorder but still keeping the RFB
elements is will make it easier to quantify how good/bad a recording is
(by measuring how many updates were delayed for example).