Change intervalThreshold
This value yielded much more accurate results on my machine. Since
sleeping only gives us a lower bound for when the thread wakes up again,
we want to tune this parameter so that it gives us a sleep time that
gets us very close to what we want, without overshooting.