Use dependency injeciton in TimedEncoder
Instead of using inheritance for each subclassed encoder, we turn
TimedEncoder into an rfb::Encoder and dependency inject which encoder it
should use. New TimedEncoders are constructed using the
constructTimedEncoder function which makes sure the correct parameters
are set.

The advantage of this is that we don't need to have one class per
encoder, which should make things less cluttered and more extensible.